Move LoggerFile to teuthology.util
